Add option to build_cleanup to fail silently

+++ fastlane/actions/build_cleanup.rb
@@ -9,16 +9,21 @@ module Fastlane
 
       # Run
       def self.run(params)
-        keychain_path = lane_context[:keychain_path] || UI.user_error!("No build keychain requiring cleanup found. Did you run build_setup before build_cleanup?")
-
+        keychain_path = lane_context[:keychain_path]
         provisioning_profile_destination = "#{Dir.home}/Library/MobileDevice/Provisioning Profiles"
-        provisioning_profile_names = lane_context[:provisioning_profile_names] || UI.user_error!("No provisioning profiles requiring cleanup found. Did you run build_setup before build_cleanup?")
+        provisioning_profile_names = lane_context[:provisioning_profile_names]
 
-        Fastlane::Actions::DeleteKeychainAction.run(keychain_path: keychain_path)
-
-        provisioning_profile_names.each do |profile_name|
-          profile_path = File.join(provisioning_profile_destination, profile_name)
-          File.delete(profile_path) if File.exist?(profile_path)
+        if keychain_path.nil? || provisioning_profile_names.nil? then
+          if params[:fail_silently] == false then
+            UI.user_error!("No build keychain or provisioning profiles requiring cleanup found. Did you run build_setup before build_cleanup?")
+          end
+        else
+          # Delete the keychain and provisioning profiles
+          Fastlane::Actions::DeleteKeychainAction.run(keychain_path: keychain_path)
+          provisioning_profile_names.each do |profile_name|
+            profile_path = File.join(provisioning_profile_destination, profile_name)
+            File.delete(profile_path) if File.exist?(profile_path)
+          end
         end
       end
 
@@ -36,7 +41,14 @@ module Fastlane
       end
 
       def self.available_options
-        [ ]
+        [
+          FastlaneCore::ConfigItem.new(
+            key: :fail_silently,
+            description: "Determines whether the cleanup action should emit an error if the build keychain or provisioning profiles can't be found",
+            default_value: false,
+            type: Boolean
+          )
+        ]
       end
     end
   end
